+++ src/index.ts
@@ -2,6 +2,7 @@ export { default as draftjsToMd } from './draftjsToMd';
 export { default as mdToDraftjs } from './mdToDraftjs';
 export type {
   DraftEntityMutability,
+  DraftjsToMdOptions,
   InlineStyleMapping,
   MarkdownDict,
   MdToDraftjsOptions,